+++ src/components/field/Field.tsx
@@ -12,10 +12,10 @@ import { CircularProgress, Tooltip } from "@mui/material";
 import Box from "@mui/material/Box/Box";
 import IconButton from "@mui/material/IconButton";
 import "react-toastify/dist/ReactToastify.min.css";
-import { ICircularObstacleStore } from "../../document/CircularObstacleStore";
-import { IConstraintStore } from "../../document/ConstraintStore";
-import { IEventMarkerStore } from "../../document/EventMarkerStore";
-import { IHolonomicWaypointStore } from "../../document/HolonomicWaypointStore";
+import { ICircularObstacleStore } from "../../document/stores/CircularObstacleStore";
+import { IConstraintStore } from "../../document/stores//ConstraintStore";
+import { IEventMarkerStore } from "../../document/stores/PplibCommandMarkerStore";
+import { IHolonomicWaypointStore } from "../../document/stores/HolonomicWaypointStore";
 import { Commands } from "../../document/tauriCommands";
 import CircularObstacleConfigPanel from "../config/CircularObstacleConfigPanel";
 import ConstraintsConfigPanel from "../config/ConstraintsConfigPanel";
