+++ src/document/stores/CircularObstacleStore.tsx
@@ -1,8 +1,10 @@
 import { Instance, getEnv, getParent, isAlive, types } from "mobx-state-tree";
-import { Env } from "./DocumentManager";
+
 import { ExpressionStore } from "./ExpressionStore";
-import { CircleObstacle } from "./spec/Traj";
-import { Expr } from "./spec/Misc";
+
+import { Env } from "../DocumentManager";
+import { CircleObstacle } from "../spec/Traj";
+import { Expr } from "../spec/Misc";
 
 export const CircularObstacleStore = types
   .model("CircularObstacleStore", {
