+++ src/document/stores/PathUIStore.ts
@@ -1,4 +1,5 @@
 import { types, getEnv, Instance } from "mobx-state-tree";
+
 import { Env } from "../DocumentManager";
 import { DifferentialSample, SwerveSample } from "../spec/Traj";
