+++ src/document/stores/PolygonObstacleStore.tsx
